actions: Move part into zerver.lib.bulk_create.
(cherry picked from commit f6a06ba6e3c9a51661659e191b4b5838867543e2)
